requirements.txt
ttkbootstrap
pyserial
